pyproject.toml
[project]
name = "teletranchan"
version = "0.1.0"
description = "Telegram-Kanal-Übersetzer mit OpenAI"
authors = [{ name = "Dein Name", email = "dein@email.de" }]
dependencies = ["openai>=1.0.0", "telethon>=1.30.0", "python-dotenv>=1.0.0"]
readme = "README.md"
requires-python = ">=3.8"

[tool.setuptools]
packages = ["teletranchan"]
package-dir = { "" = "src" }

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
